shared/requirements/requirements.txt
pyxdg==0.27
requests==2.27.1
mycroft-messagebus-client==0.9.4
